requirements.txt
fastapi==0.115.12
pydantic==2.11.4
pydantic[email]==2.11.4
asyncpg==0.30.0
uvicorn==0.34.2
sqlalchemy==2.0.40
python-dotenv==1.1.0
psycopg2-binary==2.9.10
pytz==2025.2
python-jose==3.4.0